Convenience functions updated to be const-correct
The functions rows, columns and frames are provided for quick access to
these values. The datatype returned by these functions is now
appropriate to the value stored and the functions themselves are now
const-correct.

+++ include/metadata.h
@@ -32,8 +32,8 @@ class Metadata
     void read( std::ifstream& );
     void reset();
 
-    unsigned rows() const;
-    unsigned columns() const;
+    unsigned short rows() const;
+    unsigned short columns() const;
     long frames() const;
 
     private:
